% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_pec_calc.R
\name{MosaicCalculator}
\alias{MosaicCalculator}
\title{Calculate mosaic percentage}
\usage{
MosaicCalculator(exp.nr, gender, snpm.data, deviations, outfile)
}
\arguments{
\item{exp.nr}{numerical experiment number}

\item{gender}{string denoting gender
( "Male", "Female" or "Unknown")}

\item{snpm.data}{dataframe containing snp data
( "Chr", "Position", "B Allele Freq")}

\item{deviations}{dataframe containing deviations
( "Chromosome Region",    "Event",    "Length",    "Cytoband",    precentage of CNV Overlap",
    "Probe Median",    "precentage Heterozygous",    "Probes",    "Count of Gene Symbols")}

\item{outfile}{optional path to direct pdf output to, if not set './Rplots.pdf' is used}
}
\description{
Takes a experiment number, gender, snp dataframe and events dataframe
and calculates the percentage of mosaic dna
}
