% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_printing_utils.R
\name{OutputToPdf}
\alias{OutputToPdf}
\title{OutputToPdf}
\usage{
OutputToPdf(output, cex = 0.7)
}
\arguments{
\item{output}{out put object}

\item{cex}{character expansion factor, default to 0.7}
}
\description{
Helper function to print the text output to a file
Takes a output object and prints it to the current output file.
}
