% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{process.deviation.event}
\alias{process.deviation.event}
\title{Main deviation.event processing fuction}
\usage{
process.deviation.event(x, snpm.data, correctionfactor, gender)
}
\arguments{
\item{x}{row containing the deviation event}

\item{snpm.data}{dataframe containing the BAF's}

\item{correctionfactor}{number used in proceing the event}

\item{gender}{string denoting gender
( "Male", "Female" or "Unknown")}
}
\description{
Main deviation.event processing fuction
}
