% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{process.deviation.line}
\alias{process.deviation.line}
\title{Function to rebuild deviations to usefull information}
\usage{
process.deviation.line(x, gender)
}
\arguments{
\item{x}{line containing raw the deviation event}

\item{gender}{string denoting gender
( "Male", "Female" or "Unknown")}
}
\description{
Function to rebuild deviations to usefull information
}
