% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_printing_utils.R
\name{RemoveIfExists}
\alias{RemoveIfExists}
\title{RemoveIfExists}
\usage{
RemoveIfExists(file)
}
\arguments{
\item{file}{file name including path to be removed}
}
\description{
Helper that removes file if it exsits
}
