% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mozaiekberekentool.R
\name{MosaicCalculator}
\alias{MosaicCalculator}
\title{Calculate mosaic percentage}
\usage{
MosaicCalculator(exp.nr, gender, snpm.data, deviations)
}
\arguments{
\item{exp.nr}{numbarical experiment number}

\item{gender}{string denoting gender
( "Male", "Female" or "Unknown")}

\item{snpm.data}{dataframe containing snp data
( "Chr", "Position", "B Allele Freq")}

\item{deviations}{dataframe containing deviations
( "Chromosome Region",    "Event",    "Length",    "Cytoband",    "% of CNV Overlap",
    "Probe Median",    "% Heterozygous",    "Probes",    "Count of Gene Symbols")
    @export}
}
\description{
Takes a experiment number, gender, snp dataframe and events dataframe
and calculates the percentage of mosaic dna
}
