% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{process.deviation.event}
\alias{process.deviation.event}
\title{Main deviation.event processing fuction}
\usage{
process.deviation.event(x, snpm.data, correctionfactor)
}
\description{
Main deviation.event processing fuction
}
