% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{run.demo}
\alias{run.demo}
\title{Run demo using expr number 1 and male = gender
Asks for file to be selected}
\usage{
run.demo()
}
\description{
Run demo using expr number 1 and male = gender
Asks for file to be selected
}
